Fixed a bug where selecting all units with MouseWheel, but then using Take instead of TakeAll removes 1 from the fire but gives amount selected.
Added a psuedo slider thumb under the radial item name to show which instance of combined items is selected.

Auto-Focus container scroll view on mouse enter.

Move Auto-Focus logic in Panel_Log into a help function in WidgetUtils.

Fixed a bug in Panel_Log when using the reveal button where it wouldn't focus the current area if it was not in the first 4 items in the list.